rss/requirements.txt
feedparser>=6.0
requests>=2.28
beautifulsoup4>=4.12
jinja2>=3.1
python-dotenv>=1.0
lxml>=4.9
